feat(): modularise code
