--- src/app.module.ts
@@ -1,36 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { ProductsController } from './controllers/products.controller';
-import { CategoriesController } from './controllers/categories.controller';
-import { OrdersController } from './controllers/orders.controller';
-import { CustomersController } from './controllers/customers.controller';
-import { BrandsController } from './controllers/brands.controller';
-import { ProductService } from './services/product.service';
-import { BrandsService } from './services/brands.service';
-import { CategoriesService } from './services/categories.service';
-import { CustomersService } from './services/customers.service';
-import { UsersController } from './controllers/users.controller';
-import { UsersService } from './services/users.service';
-
-@Module({
-  imports: [],
-  controllers: [
-    AppController,
-    ProductsController,
-    CategoriesController,
-    OrdersController,
-    CustomersController,
-    BrandsController,
-    UsersController,
-  ],
-  providers: [
-    AppService,
-    ProductService,
-    BrandsService,
-    CategoriesService,
-    CustomersService,
-    UsersService,
-  ],
-})
-export class AppModule {}
