+++ src/modules/orders/orders.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+
+import { OrdersController } from './controllers/orders.controller';
+
+@Module({
+  imports: [],
+  controllers: [OrdersController],
+  providers: [],
+})
+export class OrdersModule {}
