+++ src/modules/products/entities/brand.entity.ts
