+++ src/modules/products/entities/category.entity.ts
