+++ src/modules/products/entities/product.entity.ts
