+++ src/modules/users/users.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+
+import { CustomersController } from './controllers/customers.controller';
+import { CustomersService } from './services/customers.service';
+import { UsersController } from './controllers/users.controller';
+import { UsersService } from './services/users.service';
+
+@Module({
+  imports: [],
+  controllers: [CustomersController, UsersController],
+  providers: [CustomersService, UsersService],
+})
+export class UsersModule {}
